% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_standalone.R
\name{create_standalone}
\alias{create_standalone}
\title{Create a single unified standalione file}
\usage{
create_standalone(dir = "standalone", file_name = "import-standalone")
}
\arguments{
\item{dir}{The directory where the individual standalone files are stored.}

\item{file_name}{The name of the single standalone file.}
}
\value{
The function is called for its side effects only.
}
\description{
Create a single unified standalione file
}
