% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_standalone.R
\name{select_standalone}
\alias{select_standalone}
\title{Select standalone files from another repo}
\usage{
select_standalone(
  repo_spec,
  files = NULL,
  out_dir = "standalone",
  ref = NULL,
  host = NULL,
  error_call = caller_env()
)
}
\arguments{
\item{repo_spec}{A string identifying the GitHub repo in one of these forms:
\itemize{
\item Plain \code{OWNER/REPO} spec
\item Browser URL, such as \code{"https://github.com/OWNER/REPO"}
\item HTTPS Git URL, such as \code{"https://github.com/OWNER/REPO.git"}
\item SSH Git URL, such as \code{"git@github.com:OWNER/REPO.git"}
}}

\item{files}{Name of standalone file. The ⁠standalone-⁠ prefix and file
extension are optional. If omitted, will allow you to choose from the
standalone files offered by that repo. Unlike \link[usethis:use_standalone]{usethis::use_standalone}, selecting multiple standalone files from the same repo is allowed.}

\item{out_dir}{The directory where to download and store the individual "standalone" files. If the folder does not exist, it is automatically created. It defaults to the folder "standalone" in the root directory of the package found with \link[here:here]{here::here}.}

\item{ref}{The name of a branch, tag, or commit. By default, the file at
\code{path} will be copied from its current state in the repo's default branch.
This is extracted from \code{repo_spec} when user provides a URL.}

\item{host}{GitHub host to target, passed to the \code{.api_url} argument of
\code{\link[gh:gh]{gh::gh()}}. If \code{repo_spec} is a URL, \code{host} is extracted from that.

If unspecified, gh defaults to "https://api.github.com", although gh's
default can be customised by setting the GITHUB_API_URL environment
variable.

For a hypothetical GitHub Enterprise instance, either
"https://github.acme.com/api/v3" or "https://github.acme.com" is
acceptable.}

\item{error_call}{The execution environment of a currently running function,
e.g. call = caller_env(). The corresponding function call is retrieved and
mentioned in error messages as the source of the error.}
}
\value{
This function is mainly used for its side effects. That is, downloading and storing individual \code{"standalone"} files.
}
\description{
A "standalone" file implements a minimum set of functionality in such a way
that it can be copied into another package. \code{use_standalone()} makes it easy
to get such a file into your own repo.

It always overwrites an existing standalone file of the same name, making
it easy to update previously imported code.
}
