% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R, R/standalone-settoken.R
\name{set_envvar_token}
\alias{set_envvar_token}
\alias{get_envvar_token}
\title{Set or get a token from your  \code{.Renviron} file}
\usage{
set_envvar_token(
  token,
  install = FALSE,
  overwrite = FALSE,
  default = "TOKEN",
  quiet = FALSE,
  call = caller_env()
)

get_envvar_token(
  token = NULL,
  default = "TOKEN",
  message = NULL,
  pattern = NULL,
  perl = TRUE,
  call = caller_env(),
  ...
)

set_envvar_token(
  token,
  install = FALSE,
  overwrite = FALSE,
  default = "TOKEN",
  quiet = FALSE,
  call = caller_env()
)

get_envvar_token(
  token = NULL,
  default = "TOKEN",
  message = NULL,
  pattern = NULL,
  perl = TRUE,
  call = caller_env(),
  ...
)
}
\arguments{
\item{token}{A personal access token, API key, or other environmental
variable.}

\item{install}{Add your token to your \code{.Renviron} for use in future sessions.}

\item{overwrite}{If \code{TRUE}, overwrite any existing token in \code{.Renviron},}

\item{default}{Default name used for environmental variable where the token
is stored.}

\item{quiet}{If \code{TRUE}, suppress messages by setting the
\code{cli.default_handler} option to \code{\link[=suppressMessages]{suppressMessages()}}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{message}{Error message passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} if token can't be
found.}

\item{pattern}{Optional pattern passed to \code{\link[=grepl]{grepl()}} and used to validate the
stored token.}

\item{perl}{Passed to \code{\link[=grepl]{grepl()}} if pattern is supplied. Defaults to \code{TRUE}.}
}
\description{
This function is based on \code{\link[tidycensus:census_api_key]{tidycensus::census_api_key()}} by Kyle Walker.

\code{\link[=set_envvar_token]{set_envvar_token()}} can set an API key or personal access token (PAT) as a
named environment variable temporarily for the current session or saved for
future sessions. This function is based on \code{\link[tidycensus:census_api_key]{tidycensus::census_api_key()}} by
Kyle Walker with updates for more verbose messages using \code{{cli}} functions
and improved handling of edge cases, e.g. partial matching for token names.
}
\details{
\code{\link[=get_envvar_token]{get_envvar_token()}} can return an environment variable or error if the token
is missing or does not match a supplied pattern.
}
