% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settoken-package.R
\docType{package}
\name{settoken-package}
\alias{settoken}
\alias{settoken-package}
\title{settoken: Set and Get Tokens from Renviron}
\description{
Set or get a token stored an an environmental variable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/settoken}
  \item Report bugs at \url{https://github.com/elipousson/settoken/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
