% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone_import.R
\name{standalone-import}
\alias{standalone-import}
\alias{standalone_import}
\alias{standalone_import_all}
\title{Import standalone files.}
\usage{
standalone_import(
  repo_spec,
  file = NULL,
  ...,
  dest_dir = "./R",
  names_prefix = NULL
)

standalone_import_all(repo_spec, ..., dest_dir = "./R", names_prefix = NULL)
}
\arguments{
\item{repo_spec}{A string identifying the GitHub repo in one of these forms:
\itemize{
\item Plain \code{OWNER/REPO} spec
\item Browser URL, such as \code{"https://github.com/OWNER/REPO"}
\item HTTPS Git URL, such as \code{"https://github.com/OWNER/REPO.git"}
\item SSH Git URL, such as \code{"git@github.com:OWNER/REPO.git"}
}}

\item{file}{Name of standalone file. The ⁠standalone-⁠ prefix and file
extension are optional. If omitted, will allow you to choose from the
standalone files offered by that repo.}

\item{...}{elements passed to \code{\link[=standalone_urls]{standalone_urls()}}}

\item{dest_dir}{A string indicating the directory where the downloaded file
is to be saved. Tilde-expansion is performed.}

\item{names_prefix}{A prefix for the imported file. If \code{NULL} (default), the
prefix will be "standalone-" (e.g. \code{standalone-xyz.R}).}
}
\value{
The function returns \code{NULL} invisibly while saving the selected
standalone files in \code{dest-dir}.
}
\description{
\code{standalone_import()} allows the choice of individual standalone files.
}
\details{
\code{standalone-import_all()} imports all stadalone files available in the
selected rpository.
}
