% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone_paths.R
\name{standalone-paths}
\alias{standalone-paths}
\alias{standalone_paths_get}
\alias{standalone_paths_remove}
\title{Manage imported standalone files}
\usage{
standalone_paths_get(dest_dir = "./R", ..., names_prefix = NULL)

standalone_paths_remove(dest_dir = "./R", ..., names_prefix = NULL)
}
\arguments{
\item{dest_dir}{A string indicating the directory where the downloaded file
is to be saved. Tilde-expansion is performed.}

\item{...}{elements passed to \code{\link[=standalone_urls]{standalone_urls()}}}

\item{names_prefix}{A prefix for the imported file. If \code{NULL} (default), the
prefix will be \code{"standalone-"} (e.g. \code{standalone-xyz.R}).}
}
\value{
\code{standalone_paths_get()} returns the paths of the standalone files in \code{dest_dir}.

\code{standalone_paths_remove()} removes all standalone paths from \code{dest_dir}.
}
\description{
Manage imported standalone files
}
