% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone_import.R
\name{standalone_urls}
\alias{standalone_urls}
\title{Get standalone files standalone urls.}
\usage{
standalone_urls(
  repo_spec,
  ...,
  ref = NULL,
  host = NULL,
  error_call = caller_env()
)
}
\arguments{
\item{repo_spec}{A string identifying the GitHub repo in one of these forms:
\itemize{
\item Plain \code{OWNER/REPO} spec
\item Browser URL, such as \code{"https://github.com/OWNER/REPO"}
\item HTTPS Git URL, such as \code{"https://github.com/OWNER/REPO.git"}
\item SSH Git URL, such as \code{"git@github.com:OWNER/REPO.git"}
}}

\item{ref}{The name of a branch, tag, or commit. By default, the file at
\code{path} will be copied from its current state in the repo's default branch.
This is extracted from \code{repo_spec} when user provides a URL.}

\item{host}{GitHub host to target, passed to the \code{.api_url} argument of
\code{\link[gh:gh]{gh::gh()}}. If \code{repo_spec} is a URL, \code{host} is extracted from that.

If unspecified, gh defaults to "https://api.github.com", although gh's
default can be customised by setting the GITHUB_API_URL environment
variable.

For a hypothetical GitHub Enterprise instance, either
"https://github.acme.com/api/v3" or "https://github.acme.com" is
acceptable.}

\item{error_call}{Defaults to \code{caller_env()}. Useful for reporting errors.}
}
\value{
A named vector of download urls of standalone files.
}
\description{
Get standalone files standalone urls.
}
