% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standaloner-package.R
\docType{package}
\name{standaloner-package}
\alias{standaloner}
\alias{standaloner-package}
\title{standaloner: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/f-dallolio/standaloner}
  \item Report bugs at \url{https://github.com/f-dallolio/standaloner/issues}
}

}
\author{
\strong{Maintainer}: Filippo DallOlio \email{fdallolio@brocku.ca}

}
\keyword{internal}
