% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standaloner-package.R
\docType{package}
\name{standaloner-package}
\alias{standaloner}
\alias{standaloner-package}
\title{standaloner: Standalone Scripts for R Package Development}
\description{
Standalone scripts to set or get a token stored an an environmental variable, check objects, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/standaloner}
  \item Report bugs at \url{https://github.com/elipousson/standaloner/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
